+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 0.1.1
 
+* Enhancements to GandiV5::Domain
+  * .create now returns created domain (except in dry-run mode)
 * Enhancements to GandiV5::Email::Mailbox
   * .create now returns created mailbox
   * .create now checks for available slots and a valid type has been passed